EventUserFeedbackHiddenState and move button component
